+++ aegis_gym/rsl/envs/grasp_env_ros.py
@@ -1,11 +1,16 @@
 import math
+import tempfile
 import time
+from pathlib import Path
 from typing import Literal, Optional
 
+import cv2
+import numpy as np
 import torch as th
+from genesis.utils.geom import transform_by_quat
+from PIL import Image
 from rsl_rl.env import VecEnv
 from tensordict import TensorDict
-from genesis.utils.geom import transform_by_quat
 
 from .manipulator_ros import ManipulatorROS
 
@@ -71,9 +76,10 @@ class GraspEnvROS(VecEnv):
 
         # This pose is already in the Genesi's world base
         world_box_pose = th.tensor(
-            [0.631, 0.028, self.box_size[2] / 2 + 0.02, 0.0, 1.0, 0.0, 0.0],
+            # TODO(issue#98) move setup into URDF-dataset in ClearML
+            # [0.631, 0.028, self.box_size[2] / 2 + 0.02, 0.0, 1.0, 0.0, 0.0],
             # [0.557, 0.012, self.box_size[2] / 2 + 0.02, 0.0, 1.0, 0.0, 0.0],
-            # [0.576, 0.245, self.box_size[2] / 2 + 0.02, 0.0, 1.0, 0.0, 0.0],
+            [0.576, 0.245, self.box_size[2] / 2 + 0.02, 0.0, 1.0, 0.0, 0.0],
             device=self.device,
         )
         world_box_pose[2] += 0.00  # m
@@ -87,6 +93,21 @@ class GraspEnvROS(VecEnv):
             num_envs=self.num_envs,
             device=self.device,
         )
+
+        # TODO(issue#41) Unify the setup of the cameras
+        # TODO(issue#121) Unify the grasp_env and grasp_env_ros cameras names
+        match self.camera_setup:
+            case "default":
+                self._cameras = ["scene", "left", "right"]
+            case "scene_dual":
+                self._cameras = ["left", "right"]
+            case _:
+                raise ValueError(f"Unknown camera setup {self.camera_setup}")
+        self._cameras_order = {
+            "scene": 0,
+            "left": 1,
+            "right": 2,
+        }
 
         self._init_reward_functions()
         self._init_buffers()
@@ -263,24 +284,91 @@ class GraspEnvROS(VecEnv):
         # TODO implement gRPC image transportation
         raise NotImplementedError
 
-    def get_observations_vis(self, normalize: bool = True) -> th.Tensor:
-        match self.camera_setup:
-            case "default":
-                cams = ["scene", "left", "right"]
-            case "scene_dual":
-                cams = ["left", "right"]
-            case _:
-                raise ValueError(f"Unknown camera setup {self.camera_setup}")
+    def get_observations_vis(
+        self,
+        normalize: bool = True,
+        swap_tool_cameras: bool = False,
+        enable_vis_preview: bool = False,
+        enable_record_obs: bool = False,
+        record_dir: Optional[Path] = None,
+    ) -> th.Tensor:
+        # TODO(issue#41) Unify the camera setup
+        rgb_list: list[th.Tensor] = [None] * len(self._cameras)
 
-        rgb_list = [None] * len(cams)
-        for cam_id, cam_name in enumerate(cams):
+        for cam_name in self._cameras:
+            cam_id = self._cameras_order[cam_name]
+            if swap_tool_cameras:
+                cam_name_tmp = {"left": "right", "right": "left"}.get(
+                    cam_name, cam_name
+                )
+                cam_id = self._cameras_order[cam_name_tmp]
+
             rgb = self.robot.get_camera_frame(cam_name)
-            # rgb = rgb.permute(0, 3, 1, 2)[:, :3]
             if normalize:
                 rgb = th.clamp(rgb, 0.0, 255.0).div_(255.0)
             rgb_list[cam_id] = rgb
 
-        return th.cat(rgb_list, dim=1)
+        # TODO(issue#117) redesign the visualization preview
+        # TODO(issue#121) unify the code between grasp_env and grasp_env_ros
+        if enable_vis_preview or enable_record_obs:
+            preview = self._create_vis_observation_preview(
+                obs=rgb_list, normalize=normalize
+            )
+
+            if enable_vis_preview:
+                cv2.imshow("Visual observation preview", preview)
+                cv2.waitKey(1)
+
+            if enable_record_obs:
+                record_dir = (
+                    record_dir or Path(tempfile.gettempdir()) / "aegis_vis_obs_obs"
+                )
+                record_dir.mkdir(parents=True, exist_ok=True)
+                self._record_vis_observation(preview=preview, output_dir=record_dir)
+
+        return th.cat(rgb_list, dim=1).float()
+
+    def _create_vis_observation_preview(
+        self, obs: list[th.Tensor], normalize: bool
+    ) -> np.ndarray:
+
+        opencv_images: list[np.ndarray] = [None] * len(obs)
+        for cam_name in self._cameras:
+            cam_id = self._cameras_order[cam_name]
+
+            FIRST_IMG = 0
+            img = obs[cam_id][FIRST_IMG].permute(1, 2, 0).cpu().numpy()  # NCHW -> HWC
+            img = (img * 255).astype(np.uint8) if normalize else img.astype(np.uint8)
+            img = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
+
+            height, width = img.shape[:2]
+            max_side = 256
+            scale = min(max_side / width, max_side / height)
+            img = cv2.resize(
+                img,
+                (int(width * scale), int(height * scale)),
+                interpolation=cv2.INTER_AREA,
+            )
+            cv2.putText(
+                img,
+                str(cam_name),
+                org=(10, 30),
+                fontFace=cv2.FONT_HERSHEY_SIMPLEX,
+                fontScale=0.5,
+                color=(0, 255, 0),
+                thickness=2,
+            )
+            opencv_images[cam_id] = img
+
+        return np.hstack(opencv_images)
+
+    def _record_vis_observation(self, preview: np.ndarray, output_dir: Path) -> None:
+        preview = cv2.cvtColor(preview, cv2.COLOR_BGR2RGB)
+        record_step = getattr(self, "_record_step", 0)
+        fname = f"frame_{record_step:08d}.png"
+        Image.fromarray(preview).save(output_dir / fname)
+
+        self._record_step = record_step + 1
 
     def _reward_keypoints(self) -> th.Tensor:
         ee_pos = self.robot.ee_pose[:, :3]
@@ -364,35 +452,44 @@ class GraspEnvROS(VecEnv):
         step_4 = False
         success = False
 
-        for i in range(total_steps):
-            self.robot.read_state()
-            if i < total_steps / 5:  # go down
-                if step_1:
-                    continue
-                print("[GraspEnvROS][Demo] Going down to the grasp pose")
-                self.robot.go_to_goal(goal_pose)
-                step_1 = True
-            elif i < total_steps * 2 / 5:  # grasping
-                if step_2:
-                    continue
-                print("[GraspEnvROS][Demo] Grasping")
-                self.robot.gripper_close()
-                step_2 = True
-            elif i < total_steps * 3 / 5:  # lifting
-                if step_3:
-                    continue
-                print("[GraspEnvROS][Demo] Going up to the lift pose")
-                self.robot.go_to_goal(lift_pose)
-                fingers_width = self.robot.gripper_width
-                success = (fingers_width > min_width) and (fingers_width < max_width)
-                step_3 = True
-            else:  # reset
-                if step_4:
-                    continue
-                print("[GraspEnvROS][Demo] Going home")
-                self.robot._move_to_home()
-                self.robot.gripper_open()
-                step_4 = True
+        try:
+            for i in range(total_steps):
+                self.robot.read_state()
+                if i < total_steps / 5:  # go down
+                    if step_1:
+                        continue
+                    print("[GraspEnvROS][Demo] STEP 1: Going down to the grasp pose")
+                    self.robot.go_to_goal(goal_pose)
+                    step_1 = True
+                elif i < total_steps * 2 / 5:  # grasping
+                    if step_2:
+                        continue
+                    print("[GraspEnvROS][Demo] STEP 2: Grasping")
+                    self.robot.gripper_close()
+                    step_2 = True
+                elif i < total_steps * 3 / 5:  # lifting
+                    if step_3:
+                        continue
+                    print("[GraspEnvROS][Demo] STEP 3: Going up to the lift pose")
+                    self.robot.go_to_goal(lift_pose)
+                    fingers_width = self.robot.gripper_width
+                    success = (fingers_width > min_width) and (
+                        fingers_width < max_width
+                    )
+                    step_3 = True
+                else:  # reset
+                    if step_4:
+                        continue
+                    print("[GraspEnvROS][Demo] STEP 4: Going home")
+                    self.robot._move_to_home()
+                    self.robot.gripper_open()
+                    step_4 = True
+        except Exception as e:
+            print(f"[GraspEnvROS][Demo] Caught an exception: {e}")
+            success = 0.0
+            print("[GraspEnvROS][Demo] Going home")
+            self.robot._move_to_home()
+            self.robot.gripper_open()
 
         print(f"[GraspEnvROS][Demo] Grasp success: {success}")
         return float(success)