+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
+- [PR-113](https://github.com/AGH-CEAI/aegis_gym/pull/113) - Added flag `---debug-enable-vis-preview` for showing cameras preview.
+- [PR-113](https://github.com/AGH-CEAI/aegis_gym/pull/113) - Added flag `--debug-record-vis-obs` and `--debug-record-dir` arg for recording cameras preview into a given directory.
+- [PR-113](https://github.com/AGH-CEAI/aegis_gym/pull/113) - Added flag `--debug-swap-tool-cameras` for swapping RGB tool cameras sides (left <-> right).
+- [PR-113](https://github.com/AGH-CEAI/aegis_gym/pull/113) - Added flag `--episode-length-s` for overwriting the default episode time.
 - [PR-116](https://github.com/AGH-CEAI/aegis_gym/pull/103) - Added domain randomization for Genesis environment.
 - [PR-103](https://github.com/AGH-CEAI/aegis_gym/pull/103) - Added new `ConcatenatedCNNEncoder` encoder class.
 - [PR-112](https://github.com/AGH-CEAI/aegis_gym/pull/112) - Added script to batch-enqueue ClearML eval tasks from training runs.